+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Change Log
 - [ADDED] feature: storing to json and restoring of nets with pandas multiindex dataframes and series
 - [FIXED] create.py: if optional arguments are None or nan, the optional columns will not be added
 - [FIXED] add tap_dependent_impedance attributes to trafo3w instead of trafo, in create.create_transformer3w and create.create_transformer3w_from_parameters
+- [ADDED] add TDPF parameters as optional parameters for create line functions in create.py
 
 [2.11.1] - 2023-01-02
 -------------------------------